src/requirements.txt
azure-common==1.1.8
azure-mgmt-resource>=1.1.0
azure-mgmt-containerinstance==0.1.0
haikunator